System: transplant sidebar content from functions.php to Sidebar class
